+++ common/src/main/java/net/arna/jcraft/registry/JItemRegistry.java
@@ -100,10 +100,9 @@ public interface JItemRegistry {
     RegistrySupplier<Item> FOOLISH_SAND_BLOCK = register("foolish_sand_block",
             () -> new BlockItem(JBlockRegistry.FOOLISH_SAND_BLOCK.get(), settings()
             ));
-    RegistrySupplier<Item> SOUL_BLOCK = register("soul_block", () -> new BlockItem(JBlockRegistry.SOUL_BLOCK.get(), settings()
-    ));
-    RegistrySupplier<Item> METEORITE_BLOCK = register("meteorite_block", () -> new BlockItem(JBlockRegistry.METEORITE_BLOCK.get(), settings()
-    ));
+    RegistrySupplier<Item> SOUL_BLOCK = register("soul_block", () -> new BlockItem(JBlockRegistry.SOUL_BLOCK.get(), settings()));
+    RegistrySupplier<Item> METEORITE_BLOCK = register("meteorite_block", () -> new BlockItem(JBlockRegistry.METEORITE_BLOCK.get(), settings()));
+    RegistrySupplier<Item> POLISHED_METEORITE_BLOCK = register("polished_meteorite_block", () -> new BlockItem(JBlockRegistry.POLISHED_METEORITE_BLOCK.get(), settings()));
     RegistrySupplier<Item> METEORITE_IRON_ORE_BLOCK = register("meteorite_iron_ore_block",
             () -> new BlockItem(JBlockRegistry.METEORITE_IRON_ORE_BLOCK.get(), settings()));
     RegistrySupplier<Item> STELLAR_IRON_BLOCK = register("stellar_iron_block",